Remove main thread consumer which was seemingly causing issues with MainThreadDispatcher.cs. Fix namespaces, and rename Audio to CharismaAudio
