--- Resources/Prefabs/MainThreadConsumer.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 84d9ea5b0c5764264aa171e17fb2f613
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
